teardrop: controllable position, width, sticky
Clients can now be placed on "top", "bottom" or on the screen
"center". The last one is useful when creating scratchpad-like
windows, in which case it is useful to control width - which was not
possible before. By default a client will take 100% of the screen
width so it's easier to create drop-down bindings. Another new
property is sticky, mainly useful in the same scenario as center
placement.