scratchpad: initial import of the scratchpad manager
Scratchpad is a basic scratchpad manager. As in awesome 2 there are
two functions, one toggles scratch on the focused client and the other
toggles its visibility. In awesome 2 those were setscratch and
togglescratch; here they are scratchpad.set and scratchpad.toggle.

Set is called from a clientkeys key binding:
- scratchpad.set(c, width, height, sticky, screen)
Toggle is called from a globalkeys binding:
- scratchpad.toggle(screen)