rc.lua: make use of the arrange signal
Added fine grained border and floater controls. Code adds a titlebar
and border to each floating client (also covers floating layout), and
makes it ontop. In case there is only one tiled client visible or we
are in max layout the border is removed to save space. Next commit
will revert the change, because I don't intend to use it at this
moment but I'm placing it here for archival and reference.