rc.lua: use rule callback to remove titlebars selectively
Floating client that will NOT have a titlebar GETS one in callback. It
is removed in the manage signal. Rules are executed first, so we do it
like this, this is simpler than doing checks like: c.class ~=
"Xmessage" and c.class ~= ...

Arrange could seem like a better place to deal with the whole
floater-titlebar issue but it's not. It makes it impossible to remove
a titlebar, and causes some other side effects.